+++ test/unit/b2http/test_b2http.py
@@ -15,7 +15,7 @@ import socket
 from ..test_base import TestBase
 
 import apiver_deps
-from apiver_deps_exception import BadDateFormat, BadJson, BrokenPipe, B2ConnectionError, ClockSkew, ConnectionReset, ServiceError, UnknownError, UnknownHost, TooManyRequests, InvalidJsonResponse
+from apiver_deps_exception import BadDateFormat, BadJson, BrokenPipe, B2ConnectionError, ClockSkew, ConnectionReset, ServiceError, UnknownError, UnknownHost, TooManyRequests, InvalidJsonResponse, PotentialS3EndpointAsRealmPassed
 from apiver_deps import USER_AGENT
 from apiver_deps import B2Http
 from apiver_deps import B2HttpApiConfig
@@ -102,12 +102,22 @@ class TestTranslateErrors(TestBase):
         response = MagicMock()
         response.status_code = 400
         response.content = b'{' * 500
+        response.url = 'https://example.com'
 
         with self.assertRaises(InvalidJsonResponse) as error:
             B2Http._translate_errors(lambda: response)
 
             content_length = min(len(response.content), len(error.content))
             self.assertEqual(response.content[:content_length], error.content[:content_length])
+
+    def test_potential_s3_endpoint_passed_as_realm(self):
+        response = MagicMock()
+        response.status_code = 400
+        response.content = b'<?xml version="1.0" encoding="UTF-8" standalone="yes"?>'
+        response.url = 'https://s3.us-west-000.backblazeb2.com'
+
+        with self.assertRaises(PotentialS3EndpointAsRealmPassed):
+            B2Http._translate_errors(lambda: response)
 
 
 class TestTranslateAndRetry(TestBase):